+++ src/search/classic/node.cc
@@ -41,6 +41,7 @@
 #include "utils/hashcat.h"
 
 namespace lczero {
+namespace classic {
 
 /////////////////////////////////////////////////////////////////////////
 // Node garbage collector
@@ -539,4 +540,5 @@ void NodeTree::DeallocateTree() {
   current_head_ = nullptr;
 }
 
+}  // namespace classic
 }  // namespace lczero
