+++ src/search/classic/stoppers/alphazero.cc
@@ -28,6 +28,7 @@
 #include "mcts/stoppers/stoppers.h"
 
 namespace lczero {
+namespace classic {
 
 namespace {
 
@@ -73,4 +74,5 @@ std::unique_ptr<TimeManager> MakeAlphazeroTimeManager(
     int64_t move_overhead, const OptionsDict& params) {
   return std::make_unique<AlphazeroTimeManager>(move_overhead, params);
 }
+}  // namespace classic
 }  // namespace lczero
