+++ src/search/classic/stoppers/simple.cc
@@ -28,6 +28,7 @@
 #include "mcts/stoppers/stoppers.h"
 
 namespace lczero {
+namespace classic {
 
 namespace {
 
@@ -126,4 +127,5 @@ std::unique_ptr<TimeManager> MakeSimpleTimeManager(int64_t move_overhead,
                                                    const OptionsDict& params) {
   return std::make_unique<SimpleTimeManager>(move_overhead, params);
 }
+}  // namespace classic
 }  // namespace lczero
