+++ src/search/classic/stoppers/simple.h
@@ -30,8 +30,10 @@
 #include "utils/optionsdict.h"
 
 namespace lczero {
+namespace classic {
 
 std::unique_ptr<TimeManager> MakeSimpleTimeManager(
     int64_t move_overhead, const OptionsDict& params);
 
+}  // namespace classic
 }  // namespace lczero
