+++ src/search/classic/stoppers/smooth.cc
@@ -36,6 +36,7 @@
 #include "utils/mutex.h"
 
 namespace lczero {
+namespace classic {
 namespace {
 
 class Params {
@@ -636,4 +637,5 @@ std::unique_ptr<TimeManager> MakeSmoothTimeManager(int64_t move_overhead,
   return std::make_unique<SmoothTimeManager>(move_overhead, params);
 }
 
+}  // namespace classic
 }  // namespace lczero
