+++ src/search/classic/stoppers/smooth.h
@@ -31,10 +31,10 @@
 #include "utils/optionsdict.h"
 
 namespace lczero {
+namespace classic {
 
-float ComputeEstimatedMovesToGo(int ply, float midpoint, float steepness);
-
-std::unique_ptr<TimeManager> MakeLegacyTimeManager(int64_t move_overhead,
+std::unique_ptr<TimeManager> MakeSmoothTimeManager(int64_t move_overhead,
                                                    const OptionsDict& params);
 
+}  // namespace classic
 }  // namespace lczero
