+++ src/search/classic/stoppers/stoppers.h
@@ -34,6 +34,7 @@
 #include "mcts/stoppers/timemgr.h"
 
 namespace lczero {
+namespace classic {
 
 // Combines multiple stoppers into one.
 class ChainedSearchStopper : public SearchStopper {
@@ -151,4 +152,5 @@ class SmartPruningStopper : public SearchStopper {
   std::optional<int64_t> first_eval_time_ GUARDED_BY(mutex_);
 };
 
+}  // namespace classic
 }  // namespace lczero
