+++ src/search/classic/stoppers/timemgr.h
@@ -37,6 +37,7 @@
 #include "utils/optionsdict.h"
 
 namespace lczero {
+namespace classic {
 
 // Various statistics that search sends to stoppers for their stopping decision.
 // It is expected that this structure will grow.
@@ -108,4 +109,5 @@ class TimeManager {
                                                     const NodeTree& tree) = 0;
 };
 
+}  // namespace classic
 }  // namespace lczero
